pyproject.toml
[project]
name = "voice-news-agent"
version = "0.1.0"
description = "Voice-activated news agent with real-time interruption and smart memory"
authors = [
    {name = "Voice News Agent Team", email = "team@voicenewsagent.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["voice", "news", "ai", "agent", "real-time"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # FastAPI and web framework
    "fastapi==0.118.2",
    "uvicorn[standard]==0.37.0",
    "websockets==15.0.1",
    "python-multipart==0.0.20",
    # Authentication and security
    "python-jose[cryptography]==3.5.0",
    "passlib[bcrypt]==1.7.4",
    # Configuration and environment
    "python-dotenv==1.1.1",
    "pydantic-settings==2.11.0",
    # Database and ORM
    "supabase==2.22.0",
    "asyncpg==0.30.0",
    "sqlalchemy==2.0.43",
    "alembic==1.16.5",
    # Cache
    "redis==6.4.0",
    "upstash-redis==1.4.0",
    # AI and LLM
    "langchain==0.3.27",
    "langchain-openai==0.3.35",
    "openai==2.2.0",
    # Financial data
    "yfinance==0.2.66",
    "alpha-vantage==3.0.0",
    # Voice and audio (lightweight, production-ready)
    "edge-tts==7.0.2",
    "webrtcvad==2.0.10",
    "pydub==0.25.1",
    "gradio-client==1.13.3", # For HuggingFace Space API
    # Language detection
    "langid==1.1.6",
    "langdetect==1.0.9",
    # HTTP and API
    "httpx==0.28.1",
    "aiofiles==24.1.0",
    "pydantic==2.12.0",
    # Utilities
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "typing-extensions==4.15.0",
    # Logging
    "loguru>=0.7.0",
    # Audio processing
    "numpy>=1.24.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # Reddit API
    "praw>=7.8.1",
    "asyncpraw>=7.8.1",
    # LangGraph
    "langgraph>=1.0.0",
    "jupyter>=1.1.1",
    "ipykernel>=7.0.1",
    "playwright>=1.55.0",
    "rich>=14.2.0",
    "tavily>=1.1.0",
    "pypdf>=6.1.3",
]

[project.optional-dependencies]
# Local ASR dependencies (heavy - only for local development)
# Install with: uv sync --extra local-asr
local-asr = [
    "funasr>=1.0.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "sounddevice>=0.5.2",
    "pygame>=2.5.2",
]

dev = [
    # Development tools
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

test = [
    # Testing framework
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0.0",
    "pytest-xdist>=3.3.0",
    "pytest-benchmark>=4.0.0",
    "pytest-html>=3.2.0",
    "pytest-json-report>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/voicenewsagent/voice-news-agent"
Documentation = "https://docs.voicenewsagent.com"
Repository = "https://github.com/voicenewsagent/voice-news-agent"
Issues = "https://github.com/voicenewsagent/voice-news-agent/issues"

[project.scripts]
voice-news-agent = "src.main:main"
voice-news-agent-server = "backend.app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--color=yes",
    "--durations=10",
    "--timeout=15"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "backend: marks tests as backend tests",
    "src: marks tests as source component tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["backend", "src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B011"]
